Insert working
Insert is working for me after I changed the datatypes for creditnum and phone to varchars instead of ints. I did this because I think it is more likely that a query would be searching for a substring of one of these columns, rather than performing mathematical operations on them.